requirements-dev.txt
# Development and Testing Dependencies
# ======================================

# Core Django requirements (inherit from production)
-r requirements.txt

# Testing Framework
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-xdist==3.6.1  # parallel test execution

# Test Data Generation
model-bakery==1.19.5

# Time Mocking
freezegun==1.5.1

# HTTP Mocking (for Ollama and external APIs)
responses==0.25.3
httpx[http2]==0.27.2  # ensure httpx is present for mocking

# Code Quality
coverage[toml]==7.6.4
flake8==7.1.1
black==24.10.0
isort==5.13.2

# Type Checking (optional but recommended)
mypy==1.13.0
django-stubs==5.1.0

